+++ packages/core/tree/src/components/Tree/Tree-utils.js
@@ -2,20 +2,37 @@
 import type { DragState } from './Tree-types';
 import { getTreePosition } from '../../utils/tree';
 import { getDestinationPath, getSourcePath } from '../../utils/flat-tree';
-import type { Path, TreePosition, TreeData, FlattenedTree } from '../../types';
+import type {
+  Path,
+  TreeSourcePosition,
+  TreeDestinationPosition,
+  TreeData,
+  FlattenedTree,
+} from '../../types';
 
 /*
-    Translates a drag&drop movement from a purely index based flat list style to tree-friendly `TreePosition` data structure 
-    to make it available in the onDragEnd callback.  
+    Translates a drag&drop movement from an index based position to a relative (parent, index) position
 */
 export const calculateFinalDropPositions = (
   tree: TreeData,
   flattenedTree: FlattenedTree,
   dragState: DragState,
-): { sourcePosition: TreePosition, destinationPosition: ?TreePosition } => {
-  const { source, destination, horizontalLevel } = dragState;
+): {
+  sourcePosition: TreeSourcePosition,
+  destinationPosition: ?TreeDestinationPosition,
+} => {
+  const { source, destination, combine, horizontalLevel } = dragState;
   const sourcePath: Path = getSourcePath(flattenedTree, source.index);
-  const sourcePosition: TreePosition = getTreePosition(tree, sourcePath);
+  const sourcePosition: TreeSourcePosition = getTreePosition(tree, sourcePath);
+
+  if (combine) {
+    return {
+      sourcePosition,
+      destinationPosition: {
+        parentId: combine.draggableId,
+      },
+    };
+  }
 
   if (!destination) {
     return { sourcePosition, destinationPosition: null };
@@ -27,9 +44,8 @@ export const calculateFinalDropPositions = (
     destination.index,
     horizontalLevel,
   );
-  const destinationPosition: ?TreePosition = getTreePosition(
-    tree,
-    destinationPath,
-  );
+  const destinationPosition: ?TreeDestinationPosition = {
+    ...getTreePosition(tree, destinationPath),
+  };
   return { sourcePosition, destinationPosition };
 };