+++ packages/editor/editor-core/src/__tests__/integration/card/inline-2.ts
@@ -26,7 +26,7 @@ const moveUp = (page, selector) => {
   BrowserTestCase(
     `pasting an link then typing still converts to inline card`,
     {
-      skip: ['Chrome', 'chrome', 'ie', 'safari'],
+      skip: ['chrome', 'ie', 'safari'],
     },
     async client => {
       let browser = await new Page(client);