+++ packages/editor/editor-core/src/__tests__/integration/extensions/insert-extension.ts
@@ -12,7 +12,7 @@ import { messages } from '../../../plugins/insert-block/ui/ToolbarInsertBlock';
 ['Inline', 'Block'].forEach(async extensionType => {
   BrowserTestCase(
     `Extension: Insert ${extensionType} extension`,
-    { skip: ['edge', 'ie'] },
+    { skip: ['ie'] },
     async client => {
       const page = await new Page(client);
       await page.goto(fullpage.path);