+++ packages/media/media-store/src/index.ts
@@ -10,6 +10,3 @@ export {
   UploadFileCallbacks,
   UploadFileResult,
 } from './uploader';
-
-// Remove this export as soon as we deprecate oldUploadService
-export { createHasher } from './utils/hashing/hasherCreator';