+++ packages/search/global-search/src/__tests__/unit/mocks/_mockCrossProductSearchClient.ts
@@ -1,7 +1,9 @@
 import {
+  ABTest,
   CrossProductSearchClient,
   CrossProductSearchResults,
   EMPTY_CROSS_PRODUCT_SEARCH_RESPONSE,
+  SearchSession,
 } from '../../../api/CrossProductSearchClient';
 import { Scope } from '../../../api/types';
 import { Result } from '../../../model/Result';
@@ -13,17 +15,23 @@ export function makeSingleResultCrossProductSearchResponse(
 ): CrossProductSearchResults {
   const response = new Map();
   response.set(scope, [result || makeJiraObjectResult()]);
-  return { experimentId: 'experiment-1', results: response };
+  return { results: response };
 }
 
 export const noResultsCrossProductSearchClient: CrossProductSearchClient = {
   search(query: string) {
     return Promise.resolve(EMPTY_CROSS_PRODUCT_SEARCH_RESPONSE);
   },
+  getAbTestData(scope: Scope, searchSession: SearchSession) {
+    return Promise.resolve(undefined);
+  },
 };
 
 export const errorCrossProductSearchClient: CrossProductSearchClient = {
   search(query: string) {
+    return Promise.reject('error');
+  },
+  getAbTestData(scope: Scope, searchSession: SearchSession) {
     return Promise.reject('error');
   },
 };
@@ -35,5 +43,27 @@ export function singleResultCrossProductSearchClient(
     search(query: string) {
       return Promise.resolve(makeSingleResultCrossProductSearchResponse(scope));
     },
+    getAbTestData(scope: Scope, searchSession: SearchSession) {
+      return Promise.resolve(undefined);
+    },
   };
 }
+
+export const mockCrossProductSearchClient = (
+  data: CrossProductSearchResults,
+  abTest?: ABTest,
+): CrossProductSearchClient => ({
+  search(
+    query: string,
+    searchSession: SearchSession,
+    scopes: Scope[],
+  ): Promise<CrossProductSearchResults> {
+    return Promise.resolve(data);
+  },
+  getAbTestData(
+    scope: Scope,
+    searchSession: SearchSession,
+  ): Promise<ABTest | undefined> {
+    return Promise.resolve(abTest);
+  },
+});